third_party/WebKit/LayoutTests/fast/performance/performance-measure-exceptions-expected.txt

This tests that 'performance.measure' throws exceptions with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.performance.measure("DoesNotExist", "AlsoDoesNotExist") threw exception SyntaxError: Failed to execute 'measure' on 'Performance': The mark 'AlsoDoesNotExist' does not exist..
PASS window.performance.measure("measuring", 'unloadEventStart', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'unloadEventStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'unloadEventEnd', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'unloadEventEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'redirectStart', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'redirectStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'redirectEnd', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'redirectEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'secureConnectionStart', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'secureConnectionStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domInteractive', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'domInteractive' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domContentLoadedEventStart', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'domContentLoadedEventStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domContentLoadedEventEnd', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'domContentLoadedEventEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domComplete', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'domComplete' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'loadEventStart', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'loadEventStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'loadEventEnd', "mark") threw exception InvalidAccessError: Failed to execute 'measure' on 'Performance': 'loadEventEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 51.15, "mark") threw exception SyntaxError: Failed to execute 'measure' on 'Performance': The mark '51.15' does not exist..
PASS window.performance.measure("measuring", null, "mark") threw exception SyntaxError: Failed to execute 'measure' on 'Performance': The mark 'null' does not exist..
PASS window.performance.measure("measuring", {}, "mark") threw exception SyntaxError: Failed to execute 'measure' on 'Performance': The mark '[object Object]' does not exist..
PASS successfullyParsed is true

TEST COMPLETE
